Change Profile app pages structure
