+++ profiles/urls.py
@@ -4,5 +4,6 @@ from . import views
 
 app_name = "profiles"
 urlpatterns = [
-    path("", views.ProfileDetail.as_view(), name="detail"),
+    path("", views.ProfileList.as_view(), name="index"),
+    path("<int:pk>/", views.ProfileDetail.as_view(), name="detail"),
 ]
