+++ profiles/views/__init__.py
@@ -1,1 +1,2 @@
 from .profile_detail import ProfileDetail
+from .profile_list import ProfileList
