+++ profiles/views/profile_detail.py
@@ -22,7 +22,7 @@ class ProfileDetail(
 
     def get_queryset(self) -> QuerySet:
         """Profile view returns a user's profile."""
-        return models.Profile.objects.filter(user__id=self.request.user.id)
+        return models.Profile.objects.all()
 
     def get(self, request: HttpRequest, *args: Any, **kwargs: Any) -> response.Response:
         """Handle get request by return detail of a profile.
@@ -30,19 +30,7 @@ class ProfileDetail(
         :param request: Http request object
         :return: Http response object
         """
-        return self.list(request, *args, **kwargs)
-
-    def list(self, request: HttpRequest, *args: Any, **kwargs: Any) -> response.Response:
-        queryset = self.filter_queryset(self.get_queryset())
-
-        page = self.paginate_queryset(queryset)
-        if page is not None:
-            serializer = self.get_serializer(page, many=True)
-            return self.get_paginated_response(serializer.data)
-
-        serializer = self.get_serializer(queryset, many=True)
-        return response.Response(serializer.data + [{
-                "has_profile": models.Profile.has_profile(request.user)}])
+        return self.retrieve(request, *args, **kwargs)
 
     def put(self, request: HttpRequest, *args: Any, **kwargs: Any) -> response.Response:
         """Handle put request by edit a profile.
@@ -50,37 +38,13 @@ class ProfileDetail(
         :param request: Http request object
         :return: Http response object
         """
-        res = self.update(request, *args, **kwargs)
+        res = self.update(request, partial=True, *args, **kwargs)
 
         res_dict = res.data
 
         return response.Response(
             {
-                "message": f"You have successfully edited your KU Tangtee profile",
+                "message": f"You have successfully edited your KU Tangtee profile.",
                 "id": res_dict.get("id")
             }
-        )
-
-    def post(self, request: HttpRequest, *args: Any, **kwargs: Any) -> response.Response:
-        """Handle post request by creating a profile.
-
-        :param request: Http request object
-        :return: Http response object
-        """
-        return self.create(request, *args, **kwargs)
-
-    def create(self, request: HttpRequest, *args: Any, **kwargs: Any) -> response.Response:
-        """Create new profile instance.
-
-        :param request: Http request object
-        :return: Http response object
-        """
-        data = {"user": request.user.id} | request.data
-        serializer = self.get_serializer(data=data)
-        serializer.is_valid(raise_exception=True)
-        new_profile = serializer.save()
-        headers = self.get_success_headers(serializer.data)
-        return response.Response(
-            {'message': f"You have successfully created your KU Tangtee profile" , "id": new_profile.id},
-            status=status.HTTP_201_CREATED, headers=headers
         )
